Add files via upload
Added the txt files of UT antitrans legislation bills HB132(failed), HB0209, HB0257, HB0269, SB0016, SB0039, SB0093. The cleaned_blah.csv are the cleaned up versions of the text removing stop words and large blank spaces, etc. And add the ACLU csv for years 2023, 2024, and 2025(which was pulled in July, so does not have a complete picture of 2025 yet)